step4

+++ models/tweet.py
@@ -1,12 +1,12 @@
 from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy import Column, Integer, String, Date, Float
+from sqlalchemy import Column, Integer, String, Date, Float, BigInteger
 
 Base = declarative_base()
 
 class Tweet(Base):
     __tablename__ = "tweet"
     id = Column(Integer, primary_key=True, not_null=True)
-    tweet_id = Column(Integer, not_null=True)
+    tweet_id = Column(BigInteger, not_null=True)
     user_id = Column(Integer, not_null=True)
     text = Column(String, not_null=True)
     created_at = Column(Date, not_null=True)
