+++ models/tweet_timeseries.py
@@ -1,12 +1,12 @@
 from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy import Column, Integer, String, Date, Float
+from sqlalchemy import Column, Integer, String, Date, Float, BigInteger
 
 Base = declarative_base()
 
 class Tweet_Timeseries(Base):
     __tablename__ = "tweet_timeseries"
     id = Column(Integer, primary_key=True, not_null=True)
-    tweet_id = Column(Integer, not_null=True)
+    tweet_id = Column(BigInteger, not_null=True)
     retweet_count = Column(Integer, not_null=True)
     favorite_count = Column(Integer, not_null=True)
     date = Column(Date, not_null=True)
